requirements.txt
# 基础依赖
Pillow>=10.0.0  # PIL 图像处理库

# Roon 相关
roonapi>=1.0.0  # Roon API 客户端

# 驱动相关
evdev>=1.0.0  # 事件驱动设备库
gpiozero>=1.0.0  # GPIO 库
smbus>=1.0.0  # I2C 库
spidev>=1.0.0  # SPI 库
smbus2>=0.4.1  # I2C 库